requirements.txt
pandas
numpy
textblob
matplotlib
streamlit
python-dotenv
nltk
pandasai
langchain-groq
